+++ webapp/chlamdb/urls.py
@@ -9,6 +9,7 @@ favicon_view = RedirectView.as_view(url='/assets/favicon.ico', permanent=True)
 
 
 urlpatterns = [
+    re_path(r'^vf_comparison', tabular_comparison.VfComparisonView.as_view(), name="vf_comparison"),  # noqa
     re_path(r'^venn_vf/$', venn.VennVfView.as_view(), name="venn_vf"),
     re_path(r'^venn_pfam/$', venn.VennPfamView.as_view(), name="venn_pfam"),
     re_path(r'^venn_orthogroup/$', venn.VennOrthogroupView.as_view(), name="venn_orthogroup"),  # noqa