+++ webapp/views/utils.py
@@ -31,7 +31,7 @@ title2page = {
         'orthogroup_comparison', 'pan_genome_orthogroup', 'plot_heatmap_orthogroup',
         'venn_orthogroup'],
     'Comparisons: Virulence Factors': [
-        'entry_list_vf', 'index_comp_vf', 'extract_vf', 'venn_vf'],
+        'entry_list_vf', 'index_comp_vf', 'extract_vf', 'venn_vf', 'vf_comparison'],
     'Genome alignments: Circos plot': ['circos'],
     'Genome alignments: Plot region': ['plot_region'],
     'Genomes: table of contents': ['extract_contigs', 'genomes_intro'],